requirements.txt
absl-py==1.4.0
accelerate==0.21.0
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
asttokens==2.2.1
backcall==0.2.0
beautifulsoup4==4.12.2
brotlipy==0.7.0
cachetools==5.3.1
cfgv==3.3.1
click==8.1.6
contourpy==1.1.0
cycler==0.11.0
decorator==4.4.2
diffusers==0.19.3
distlib==0.3.7
docker-pycreds==0.4.0
docstring-parser==0.15
einops==0.6.1
ema-pytorch==0.2.3
executing==1.2.0
filelock==3.12.2
fonttools==4.42.0
frozendict==2.3.8
fsspec==2023.6.0
gdown==4.7.1
gitdb==4.0.10
GitPython==3.1.32
google-auth==2.22.0
google-auth-oauthlib==1.0.0
grpcio==1.57.0
huggingface-hub==0.16.4
hydra-core==1.3.2
identify==2.5.26
imageio==2.31.1
imageio-ffmpeg==0.4.8
importlib-metadata==6.8.0
importlib-resources==6.0.1
ipython==8.14.0
jaxtyping==0.2.20
jedi==0.19.0
kiwisolver==1.4.4
kornia==0.6.0
lazy_loader==0.3
line-profiler==4.0.3
lpips==0.1.4
Markdown==3.4.4
markdown-it-py==3.0.0
matplotlib==3.7.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mediapy==1.1.9
moviepy==1.0.3
msgpack==1.0.5
munch==4.0.0
nodeenv==1.8.0
numpy==1.25.2
nvidia-ml-py==12.535.77
nvitop==1.2.0
oauthlib==3.2.2
omegaconf==2.3.0
opencv-python==4.8.0.76
packaging==23.1
parso==0.8.3
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==3.10.0
pre-commit==3.3.2
proglog==0.1.10
prompt-toolkit==3.0.39
protobuf==4.24.0
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
Pygments==2.16.1
pyliblzfse==0.4.1
pyparsing==3.0.9
python-dateutil==2.8.2
PyWavelets==1.4.1
PyYAML==6.0.1
regex==2023.8.8
requests-oauthlib==1.3.1
rich==13.5.2
rsa==4.9
safetensors==0.3.2
scikit-image==0.21.0
scipy==1.11.1
sentry-sdk==1.29.2
setproctitle==1.3.2
shtab==1.6.4
six==1.16.0
smmap==5.0.0
soupsieve==2.4.1
stack-data==0.6.2
tensorboard==2.14.0
tensorboard-data-server==0.7.1
termcolor==2.3.0
tifffile==2023.7.18
tokenizers==0.13.3
torch==2.0.1
torchaudio==2.0.2
torchtyping==0.1.4
torchvision==0.15.2
tqdm==4.66.1
traitlets==5.9.0
transformers==4.31.0
trimesh==3.23.1
triton==2.0.0
typeguard==4.1.0
tyro==0.5.5
virtualenv==20.24.2
viser==0.0.16
wandb==0.15.8
wcwidth==0.2.6
websockets==11.0.3
Werkzeug==2.3.6
zipp==3.16.2
